pyproject.toml
[project]
name = "dspy-reference-examples"
version = "0.1.0"
description = "DSPy optimization reference for Ozempic complaint classification"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "dspy-ai>=2.5.0",
  "openai>=1.0.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.4.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = [
  "integration: tests that hit real LLMs or external services",
]
log_cli = true
log_cli_level = "INFO"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
]
ignore = [
  "E501",  # Line too long - often not auto-fixable, especially in strings
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
